+++ tests/slist_test.cc
@@ -160,3 +160,36 @@ TEST(SList, foreach) {
 
   slist_deinit(s);
 }
+
+TEST(SList, insert) {
+  string log;
+  slist_t slist;
+  slist_t* s = &slist;
+
+  slist_init(s, NULL, NULL);
+
+  ASSERT_EQ(slist_size(s), 0);
+  ASSERT_EQ(slist_insert(s, 0, TO_POINTER(0)), RET_OK);
+  ASSERT_EQ(slist_insert(s, 1, TO_POINTER(2)), RET_OK);
+  ASSERT_EQ(slist_insert(s, 1, TO_POINTER(1)), RET_OK);
+  log="";
+  slist_foreach(s, visit_dump, &log);
+  ASSERT_EQ(log, "0:1:2:");
+
+  ASSERT_EQ(slist_insert(s, 0, TO_POINTER(5)), RET_OK);
+  log="";
+  slist_foreach(s, visit_dump, &log);
+  ASSERT_EQ(log, "5:0:1:2:");
+  
+  ASSERT_EQ(slist_insert(s, 3, TO_POINTER(6)), RET_OK);
+  log="";
+  slist_foreach(s, visit_dump, &log);
+  ASSERT_EQ(log, "5:0:1:6:2:");
+  
+  ASSERT_EQ(slist_insert(s, 10, TO_POINTER(10)), RET_OK);
+  log="";
+  slist_foreach(s, visit_dump, &log);
+  ASSERT_EQ(log, "5:0:1:6:2:10:");
+
+  slist_deinit(s);
+}